+++ birdnet-pi/rootfs/etc/cont-init.d/31-checks.sh
@@ -58,5 +58,7 @@ sed -i "s|/etc/birdnet/birdnet.conf|/config/birdnet.conf|g" "$HOME"/BirdNET-Pi/s
 sed -i "/update_caddyfile/c echo \"yes\"" "$HOME"/BirdNET-Pi/scripts/update_birdnet_snippets.sh # Avoid systemctl
 
 # Execute update_birdnet_snippets
-chmod +777 "$HOME"/BirdNET-Pi/scripts/update_birdnet_snippets.sh
+export RECS_DIR="$HOME/BirdSongs"
+export EXTRACTED="$HOME/BirdSongs/Extracted"
+chmod +x "$HOME"/BirdNET-Pi/scripts/update_birdnet_snippets.sh
 "$HOME"/BirdNET-Pi/scripts/update_birdnet_snippets.sh