+++ birdnet-pi/rootfs/etc/cont-init.d/81-modifications.sh
@@ -64,7 +64,7 @@ sed -i "s|/stats|/stats/|g" "$HOME/BirdNET-Pi/homepage/views.php"
 sed -i "s|/log|/log/|g" "$HOME/BirdNET-Pi/homepage/views.php"
 
 # Check if port 80 is correctly configured
-if [ -n "$(bashio::addon.port 80)" ] && [ "$(bashio::addon.port 80)" != 80 ]; then
+if [ -n "$(bashio::addon.port "80")" ] && [ "$(bashio::addon.port "80")" != 80 ]; then
     bashio::log.fatal "The port 80 is enabled, but should still be 80 if you want automatic SSL certificates generation to work."
 fi
 
@@ -96,6 +96,9 @@ sed -i -e '/<option disabled selected>/s/selected//' \
 
 # Correct language labels according to birdnet.conf
 echo "... adapting labels according to birdnet.conf"
-export "$(grep "^DATABASE_LANG" /config/birdnet.conf)" || bashio::log.warning "DATABASE_LANG not found in configuration. Using default labels."
-bashio::log.info "Setting language to $DATABASE_LANG"
-"$HOME/BirdNET-Pi/scripts/install_language_label_nm.sh" -l "$DATABASE_LANG" &>/dev/null || bashio::log.warning "Failed to update language labels"
+if export "$(grep "^DATABASE_LANG" /config/birdnet.conf)"; then
+    bashio::log.info "Setting language to ${DATABASE_LANG:-en}"
+else
+    bashio::log.warning "DATABASE_LANG not found in configuration. Using default labels."
+fi
+"$HOME/BirdNET-Pi/scripts/install_language_label_nm.sh" -l "${DATABASE_LANG:-}" &>/dev/null || bashio::log.warning "Failed to update language labels"