+++ netalertx/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 25.2.8-3 (09-02-2025)
+- Minor bugs fixed
+## 25.2.8-2 (09-02-2025)
+- Minor bugs fixed
 
 ## 25.2.8 (08-02-2025)
 - Update to latest version from jokob-sk/NetAlertX (changelog : https://github.com/jokob-sk/NetAlertX/releases)